pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "archive>=0.3",
    "mcp>=1.10.1",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "nest-asyncio>=1.5.0",
]
